feat(cream-ksp): generate per-child @CopyToChildren funName with leaf-count validation
Resolve the @CopyToChildren `funName` template per concrete leaf so each
generated function name stays unique. A per-target token (e.g.
CopyTargetSimpleName) is required when more than one function is generated;
a plain-literal funName is rejected in that case.

The "more than one function" decision uses the real generated leaf count —
flattening transitive sealed nodes and respecting notCopyToObject (object
leaves dropped) — so a single-effective-leaf hierarchy still accepts a literal
name. The shared transitive-leaf flattener lives in util/ksp
(SealedHierarchy) as a generic KSP helper — the SSoT for both @SealedCopy and
@CopyToChildren.
