+++ cream-ksp/src/main/kotlin/me/tbsten/cream/ksp/core/sealedCopy/SealedCopy.kt
@@ -10,6 +10,7 @@ import me.tbsten.cream.ksp.core.common.GenerateSourceAnnotation
 import me.tbsten.cream.ksp.core.common.annotationsOf
 import me.tbsten.cream.ksp.core.common.toModifierString
 import me.tbsten.cream.ksp.util.ksp.asString
+import me.tbsten.cream.ksp.util.ksp.collectConcreteSubclasses
 import java.io.BufferedWriter
 
 /**
